refactor: Переместить в другие пакеты
Создал в основных пакетах новые пакеты сompany и insurance. Переместил туда классы, связанные с этим критерием.
