+++ src/main/java/ru/ya/insurance/dto/insurance/InsuranceOfferCardDto.java
@@ -1,8 +1,9 @@
-package ru.ya.insurance.dto.common;
+package ru.ya.insurance.dto.insurance;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
+import ru.ya.insurance.dto.insurance.FeatureDto;
 
 import java.util.List;
