+++ src/main/java/ru/ya/insurance/mapper/insurance/InsuranceOfferCardMapper.java
@@ -1,12 +1,16 @@
-package ru.ya.insurance.mapper.common;
+package ru.ya.insurance.mapper.insurance;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import org.mapstruct.ReportingPolicy;
-import ru.ya.insurance.dto.common.FeatureDto;
-import ru.ya.insurance.dto.common.InsuranceOfferCardDto;
-import ru.ya.insurance.model.common.*;
+import ru.ya.insurance.dto.insurance.FeatureDto;
+import ru.ya.insurance.dto.insurance.InsuranceOfferCardDto;
+import ru.ya.insurance.model.company.License;
+import ru.ya.insurance.model.insurance.Feature;
+import ru.ya.insurance.model.insurance.Insurance;
+import ru.ya.insurance.model.insurance.RequiredDocument;
+import ru.ya.insurance.model.insurance.ValidityPeriod;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING, unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface InsuranceOfferCardMapper {
