+++ src/main/java/ru/ya/insurance/repository/insurance/InsuranceRepository.java
@@ -1,9 +1,9 @@
-package ru.ya.insurance.repository.common;
+package ru.ya.insurance.repository.insurance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import ru.ya.insurance.model.common.Insurance;
+import ru.ya.insurance.model.insurance.Insurance;
 
 import java.util.Optional;
